+++ Source/Core/DiscIO/VolumeWad.cpp
@@ -103,7 +103,7 @@ bool CVolumeWAD::GetTitleID(u8* _pBuffer) const
 	return true;
 }
 
-int CVolumeWAD::GetRevision() const
+u16 CVolumeWAD::GetRevision() const
 {
 	u16 revision;
 	if (!m_pReader->Read(m_tmd_offset + 0x1dc, 2, (u8*)&revision))
